added CDE

+++ src/tempor/models/ts_model.py
@@ -324,7 +324,7 @@ class TimeSeriesModel(nn.Module):
     ) -> float:
         y_pred = self.predict(static_data, temporal_data, observation_times)
         if self.task_type == "classification":
-            return np.mean(y_pred == outcome)
+            return np.mean(y_pred.astype(int) == outcome.astype(int))
         else:
             return np.mean(np.inner(outcome - y_pred, outcome - y_pred) / 2.0)
